back-end/pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["KossKokos <koskokos0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.0"
fastapi = "0.95.0"
sqlalchemy = "2.0.7"
uvicorn = "0.21.1"
alembic = "1.10.2"
python-dotenv = "1.0.0"
psycopg2 = "2.9.5"
psycopg2-binary = "2.9.9"
redis = "4.5.4"
python-jose = "3.3.0"
passlib = "1.7.4"
email-validator = "1.3.1"
bcrypt = "4.0.1"
python-multipart = "0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
